Fix tests.api.java.io.SerializationStressTest3.test_18_116_writeObject.
The dead field 'minExponentDigits' was of type 'byte', not 'int'.

Also add detail to serialization exception messages (this isn't just test code ---
ObjectInputStream calls this).

+++ libcore/text/src/main/java/java/text/DecimalFormat.java
@@ -1230,7 +1230,7 @@ public class DecimalFormat extends NumberFormat {
         fields.put("parseBigDecimal", parseBigDecimal);
         fields.put("symbols", symbols);
         fields.put("useExponentialNotation", false);
-        fields.put("minExponentDigits", 0);
+        fields.put("minExponentDigits", (byte) 0);
         fields.put("maximumIntegerDigits", dform.getMaximumIntegerDigits());
         fields.put("minimumIntegerDigits", dform.getMinimumIntegerDigits());
         fields.put("maximumFractionDigits", dform.getMaximumFractionDigits());
